+++ app/src/types/index.ts
@@ -16,5 +16,5 @@ export * from './tsHelper';
 export * from './whiteboard';
 export * from './actions';
 export * from './event';
-export * from './protocol';
+export * from './meetingNotes';
 export * from './legalVote';